# backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/login": {
            "post": {
                "description": "登录\n填入用户名和密码",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "用户"
                ],
                "summary": "登录 ccf",
                "parameters": [
                    {
                        "description": "data",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/response.LoginQ"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"status\":200,\"msg\":\"登录成功\",\"token\": token,\"ID\": user.UserID}",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "{\"status\":400,\"msg\":\"用户名不存在\"}",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "401": {
                        "description": "{\"status\":401,\"msg\":\"密码错误\"}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/register": {
            "post": {
                "description": "注册\n填入用户名和密码注册",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "用户"
                ],
                "summary": "注册 ccf",
                "parameters": [
                    {
                        "description": "data",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/response.RegisterQ"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"status\":200,\"msg\":\"注册成功\"}",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "{\"status\":400,\"msg\":\"用户名已存在\"}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
        "/user/headshot": {
            "post": {
                "description": "上传用户头像",
                "tags": [
                    "用户"
                ],
                "summary": "上传用户头像 ccf",
                "parameters": [
                    {
                        "type": "string",
                        "description": "用户ID",
                        "name": "user_id",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "file",
                        "description": "新头像",
                        "name": "Headshot",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"status\":200,\"msg\":\"修改成功\",\"data\":{object}}",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "{\"status\":400,\"msg\":\"用户ID不存在\"}",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "401": {
                        "description": "{\"status\":401,\"msg\":\"头像文件上传失败\"}",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "402": {
                        "description": "{\"status\":402,\"msg\":\"文件保存失败\"}",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "403": {
                        "description": "{\"status\":403,\"msg\":\"保存文件路径到数据库中失败\"}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/user/history": {
            "post": {
                "description": "获取用户浏览历史",
                "tags": [
                    "用户"
                ],
                "summary": "txc",
                "parameters": [
                    {
                        "type": "string",
                        "description": "token",
                        "name": "token",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "data",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/response.GetBrowseHistoryQ"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"status\": 200, \"msg\": \"获取成功\", \"data\": {object}}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/user/info": {
            "get": {
                "description": "查看用户个人信息",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "用户"
                ],
                "summary": "查看用户个人信息 ccf",
                "parameters": [
                    {
                        "type": "string",
                        "description": "user_id",
                        "name": "user_id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"status\":200,\"msg\":\"获取用户信息成功\",\"data\":{object}}",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "{\"status\":400,\"msg\":\"用户ID不存在\"}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/user/mod": {
            "post": {
                "description": "编辑用户信息",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "用户"
                ],
                "summary": "编辑用户信息 ccf",
                "parameters": [
                    {
                        "description": "data",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/response.ModifyQ"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"status\":200,\"msg\":\"修改个人信息成功\",\"data\":{object}}",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "{\"status\":400,\"msg\":\"用户ID不存在\"}",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "401": {
                        "description": "{\"status\":401,\"msg\":err.Error()}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/user/pwd": {
            "post": {
                "description": "编辑用户密码",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "用户"
                ],
                "summary": "编辑用户密码 ccf",
                "parameters": [
                    {
                        "description": "data",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/response.PwdModifyQ"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"status\":200,\"msg\":\"修改密码成功\",\"data\":{object}}",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "{\"status\":400,\"msg\":\"用户ID不存在\"}",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "401": {
                        "description": "{\"status\":401,\"msg\":\"原密码输入错误\"}",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "402": {
                        "description": "{\"status\":402,\"msg\":err1.Error()}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "response.GetBrowseHistoryQ": {
            "type": "object",
            "properties": {
                "page": {
                    "type": "integer"
                },
                "size": {
                    "type": "integer"
                }
            }
        },
...
        "response.LoginQ": {
            "type": "object",
            "required": [
                "password",
                "username"
            ],
            "properties": {
                "password": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                }
            }
        },
...
        "response.ModifyQ": {
            "type": "object",
            "required": [
                "user_id"
            ],
            "properties": {
                "email": {
                    "description": "邮箱",
                    "type": "string"
                },
                "fields": {
                    "description": "研究领域",
                    "type": "string"
                },
                "name": {
                    "description": "真实姓名",
                    "type": "string"
                },
                "phone": {
                    "description": "电话号码",
                    "type": "string"
                },
                "user_id": {
                    "type": "string"
                },
                "user_info": {
                    "description": "个性签名",
                    "type": "string"
                }
            }
        },
...
        "response.PwdModifyQ": {
            "type": "object",
            "required": [
                "password_new",
                "password_old",
                "user_id"
            ],
            "properties": {
                "password_new": {
                    "type": "string",
                    "minLength": 6
                },
                "password_old": {
                    "type": "string",
                    "minLength": 6
                },
                "user_id": {
                    "type": "string"
                }
            }
        },
        "response.RegisterQ": {
            "type": "object",
            "required": [
                "password",
                "username"
            ],
            "properties": {
                "password": {
                    "type": "string",
                    "minLength": 6
                },
                "username": {
                    "type": "string",
                    "minLength": 1
                }
            }
        },
...
    }
}